+++ mlkit-barcode-scanning/src/main/java/com/king/mlkit/vision/barcode/BarcodeCameraScanFragment.java
@@ -35,7 +35,7 @@ public abstract class BarcodeCameraScanFragment extends BaseCameraScanFragment<L
     /**
      * 创建分析器，默认分析所有条码格式
      *
-     * @return
+     * @return {@link Analyzer}
      */
     @Nullable
     @Override