+++ mlkit-camera-core/src/main/java/com/king/mlkit/vision/camera/BaseCameraScanFragment.java
@@ -38,7 +38,7 @@ import androidx.fragment.app.Fragment;
  * 1、通过继承 {@link BaseCameraScanActivity}或者{@link BaseCameraScanFragment}或其子类，可快速实现扫描识别。
  * （适用于大多数场景，自定义布局时需覆写getLayoutId方法）
  * <p>
- * 2、在你项目的Activity或者Fragment中实例化一个{@link BaseCameraScan}。（适用于想在扫码界面写交互逻辑，又因为项目
+ * 2、在你项目的Activity或者Fragment中实例化一个{@link BaseCameraScan}。（适用于想在扫描界面写交互逻辑，又因为项目
  * 架构或其它原因，无法直接或间接继承{@link BaseCameraScanActivity}或{@link BaseCameraScanFragment}时使用）
  * <p>
  * 3、继承{@link CameraScan}自己实现一个，可参照默认实现类{@link BaseCameraScan}，其他步骤同方式2。（高级用法，谨慎使用）
@@ -47,11 +47,22 @@ import androidx.fragment.app.Fragment;
  */
 public abstract class BaseCameraScanFragment<T> extends Fragment implements CameraScan.OnScanResultCallback<T> {
 
+    /**
+     * 相机权限请求代码
+     */
     private static final int CAMERA_PERMISSION_REQUEST_CODE = 0X86;
 
+    /**
+     * 根视图
+     */
     private View mRootView;
-
+    /**
+     * 预览视图
+     */
     protected PreviewView previewView;
+    /**
+     * CameraScan
+     */
     private CameraScan<T> mCameraScan;
 
     @Override
@@ -59,8 +70,13 @@ public abstract class BaseCameraScanFragment<T> extends Fragment implements Came
         if (isContentView()) {
             mRootView = createRootView(inflater, container);
         }
-        initUI();
         return mRootView;
+    }
+
+    @Override
+    public void onViewCreated(@NonNull View view, @Nullable Bundle savedInstanceState) {
+        super.onViewCreated(view, savedInstanceState);
+        initUI();
     }
 
     /**
@@ -68,17 +84,28 @@ public abstract class BaseCameraScanFragment<T> extends Fragment implements Came
      */
     public void initUI() {
         previewView = mRootView.findViewById(getPreviewViewId());
-        initCameraScan();
+        mCameraScan = createCameraScan(previewView);
+        initCameraScan(mCameraScan);
         startCamera();
     }
 
     /**
      * 初始化CameraScan
      */
-    public void initCameraScan() {
-        mCameraScan = createCameraScan(previewView)
-                .setAnalyzer(createAnalyzer())
+    public void initCameraScan(@NonNull CameraScan<T> cameraScan) {
+        cameraScan.setAnalyzer(createAnalyzer())
                 .setOnScanResultCallback(this);
+        initCameraScan();
+    }
+
+    /**
+     * 初始化CameraScan
+     *
+     * @Deprecated 此方法已废弃，后续可能移除；请使用 {@link #initCameraScan(CameraScan)}
+     */
+    @Deprecated
+    public void initCameraScan() {
+
     }
 
     /**
@@ -115,8 +142,8 @@ public abstract class BaseCameraScanFragment<T> extends Fragment implements Came
     /**
      * 请求Camera权限回调结果
      *
-     * @param permissions
-     * @param grantResults
+     * @param permissions 权限
+     * @param grantResults 授权结果
      */
     public void requestCameraPermissionResult(@NonNull String[] permissions, @NonNull int[] grantResults) {
         if (PermissionUtils.requestPermissionsResult(Manifest.permission.CAMERA, permissions, grantResults)) {
@@ -127,9 +154,9 @@ public abstract class BaseCameraScanFragment<T> extends Fragment implements Came
     }
 
     @Override
-    public void onDestroy() {
+    public void onDestroyView() {
         releaseCamera();
-        super.onDestroy();
+        super.onDestroyView();
     }
 
     /**
@@ -144,9 +171,9 @@ public abstract class BaseCameraScanFragment<T> extends Fragment implements Came
     /**
      * 创建{@link #mRootView}
      *
-     * @param inflater
-     * @param container
-     * @return
+     * @param inflater  {@link LayoutInflater}
+     * @param container {@link ViewGroup}
+     * @return 返回创建的根视图
      */
     @NonNull
     public View createRootView(LayoutInflater inflater, ViewGroup container) {
@@ -156,23 +183,23 @@ public abstract class BaseCameraScanFragment<T> extends Fragment implements Came
     /**
      * 布局ID；通过覆写此方法可以自定义布局
      *
-     * @return
+     * @return 布局ID
      */
     public int getLayoutId() {
         return R.layout.ml_camera_scan;
     }
 
     /**
-     * 预览界面{@link #previewView}的ID
+     * 预览视图{@link #previewView}的ID
      *
-     * @return
+     * @return 预览视图ID
      */
     public int getPreviewViewId() {
         return R.id.previewView;
     }
 
     /**
-     * 获取 {@link CameraScan}
+     * 获取{@link CameraScan}
      *
      * @return {@link #mCameraScan}
      */
@@ -183,7 +210,7 @@ public abstract class BaseCameraScanFragment<T> extends Fragment implements Came
     /**
      * 获取根视图
      *
-     * @return
+     * @return {@link #mRootView}
      */
     public View getRootView() {
         return mRootView;
@@ -193,8 +220,9 @@ public abstract class BaseCameraScanFragment<T> extends Fragment implements Came
      * 创建{@link CameraScan}
      *
      * @param previewView
-     * @return
+     * @return {@link CameraScan}
      */
+    @NonNull
     public CameraScan<T> createCameraScan(PreviewView previewView) {
         return new BaseCameraScan<>(this, previewView);
     }
@@ -202,7 +230,7 @@ public abstract class BaseCameraScanFragment<T> extends Fragment implements Came
     /**
      * 创建分析器
      *
-     * @return
+     * @return {@link Analyzer}
      */
     @Nullable
     public abstract Analyzer<T> createAnalyzer();