+++ mlkit-camera-core/src/main/java/com/king/mlkit/vision/camera/manager/BeepManager.java
@@ -22,13 +22,14 @@ import android.os.Build;
 import android.os.VibrationEffect;
 import android.os.Vibrator;
 
-
 import com.king.mlkit.vision.camera.R;
 import com.king.mlkit.vision.camera.util.LogUtils;
 
 import java.io.Closeable;
 
 /**
+ * 蜂鸣音效管理器：主要用于播放蜂鸣提示音和振动效果
+ *
  * @author <a href="mailto:jenly1314@gmail.com">Jenly</a>
  */
 public final class BeepManager implements MediaPlayer.OnErrorListener, Closeable {
@@ -47,11 +48,11 @@ public final class BeepManager implements MediaPlayer.OnErrorListener, Closeable
         updatePrefs();
     }
 
-    public void setVibrate(boolean vibrate){
+    public void setVibrate(boolean vibrate) {
         this.vibrate = vibrate;
     }
 
-    public void setPlayBeep(boolean playBeep){
+    public void setPlayBeep(boolean playBeep) {
         this.playBeep = playBeep;
     }
 
@@ -59,8 +60,8 @@ public final class BeepManager implements MediaPlayer.OnErrorListener, Closeable
         if (mediaPlayer == null) {
             mediaPlayer = buildMediaPlayer(context);
         }
-        if(vibrator == null){
-            vibrator = (Vibrator)context.getSystemService(Context.VIBRATOR_SERVICE);
+        if (vibrator == null) {
+            vibrator = (Vibrator) context.getSystemService(Context.VIBRATOR_SERVICE);
         }
     }
 
@@ -102,12 +103,12 @@ public final class BeepManager implements MediaPlayer.OnErrorListener, Closeable
 
     @Override
     public synchronized void close() {
-        try{
+        try {
             if (mediaPlayer != null) {
                 mediaPlayer.release();
                 mediaPlayer = null;
             }
-        }catch (Exception e){
+        } catch (Exception e) {
             LogUtils.e(e);
         }
     }